fix | 플레이어 체력 표시, 스테이지 넘어갈때 표시 이상 정상화

+++ Game/FragileFall/Script/HUDController.h
@@ -37,7 +37,7 @@ namespace game
         int CalcHalfHPFromPlayer() const;
         int CalcVisibleHeartCountFromPlayer() const;
         int CalcHpPerHeart() const;
-        int CalcHpPerHalfHeart() const;
+        float CalcHpPerHalfHeart() const;
 
         void ApplyHearts();
